requirements.txt
Pillow==10.1.0
Pillow==10.2.0
streamlit==1.29.0
